Move keyboard tracking to CConnection
This is a general thing so move it in to the core library instead,
letting vncviewer focus on just translation of system events to VNC
ones.
